ingress: Populate tls.secretName dynamically
This change replaces the use of `sslSecretName` in the Ingress with a
helper template that dynamically renders `<host>-tls` as the name.
Setting `tlsSecretNameOverride` will override this behavior and replaces
the secret name with an arbitrary template string instead (e.g., `foo`
or `.Values.host`).

The `-tls` suffix was chosen as the default because it's a common
pattern:

* <https://cert-manager.io/docs/usage/certificate/>
* <https://kubernetes.github.io/ingress-nginx/user-guide/tls/#default-ssl-certificate>