ingress: Remove comment in values.yaml
It's not quite true that the values will have no effect.
If `ingress.enabled=true`, the Ingress object will be created no matter
the platform.
Whether or not an Ingress object has any effect depends on if the
platform has an Ingress controller installed, so let's leave it up to
the user to decide if that's desirable or not.