web: Fix incorrect volume name
The `name` attribute in `volumeMounts` must match the name of the
volume, which in this pod is the hard-coded string `shared-volume`.
Since the default value of `.Values.persistence.name` happens to also be
`shared-volume`, this configuration did work. But in the event that a
user wanted to change `.Values.persistence.name` (which is intended to
be the name of the PVC), things would break.