fix: Include `name` key in example of imagePullSecrets
When `.Values.image.imagePullSecrets` was introduced in #174, we opted
to use the same format for `imagePullSecrets` list as used in
Kubernetes' PodSpec; it's a list of objects that requires each list
element to have the key `name`.
We failed to indicate this in the example given in the comment.
This change fixes that.