Enable configuration of more Deployment.spec field. (#207)
This change introduces the field `deploymentSpec` for web,
worker, worker-beat and flower, enabling the user to customize the spec
fields for `minReadySeconds`, `paused`, `progressDeadlineSeconds`,
`revisionHistoryLimit` and `strategy`.

If the user does not set the new values, they are not set in the
Deployments either (i.e. defaulting to Kubernetes' default value for
each field).